Add files via upload
Sure! This code defines a function called 'check' that tells you if a number is even or odd. If you pass a number to the function, it will print "number is even" if it's even and "number is odd" if it's odd. When you use check(33), it prints "number is odd" because 33 is an odd number.