Add files via upload
This code defines a function called `checks` that examines a number. If the number is positive, it prints "number is positive". If it's negative, it prints "number is negative". If it's neither positive nor negative (in other words, it's 0), it prints "number is zero". Then, it calls the `checks` function with the number 10, so it prints "number is positive".