Add files via upload
This code defines a function called `checkk` that takes a number (`num1`) as input. It checks if the number is divisible by 2, 3, both, or neither, and then prints a corresponding message. If the number is divisible by both 2 and 3, it prints "number is divisible by 2 and 3". If it's only divisible by 2, it prints "number is divisible by 2". If it's only divisible by 3, it prints "number is divisible by 3". And if it's not divisible by either, it prints "number is not divisible". Finally, it calls the function `checkk` with the number 12 as an argument.